grape_hindi_tts/requirements.txt
# Core PyTorch and audio processing
torch>=2.1.0
torchaudio>=2.1.0

# Numerical and scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Audio processing libraries
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Logging and monitoring
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.65.0

# Neural network utilities
einops>=0.7.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
pandas>=2.0.0

# Hindi language processing
indic-nlp-library>=0.92
aksharamukha>=2.0.0

# Evaluation metrics
pesq>=0.0.4
pystoi>=0.3.3
resemblyzer>=0.1.3
jiwer>=3.0.0
openai-whisper>=20230918
